Feedback: simplifying by using inheritance on the parsers instead of composition.
