+++ modules/extensions-mule-support/src/main/java/org/mule/runtime/module/extension/mule/internal/loader/parser/MuleSdkApplicationExtensionModelParser.java
@@ -4,33 +4,45 @@
  * license, a copy of which has been included with this distribution in the
  * LICENSE.txt file.
  */
-package org.mule.runtime.module.extension.mule.internal.loader.parser.metadata;
+package org.mule.runtime.module.extension.mule.internal.loader.parser;
 
 import static org.mule.runtime.api.meta.Category.COMMUNITY;
 import static org.mule.runtime.internal.dsl.DslConstants.THIS_NAMESPACE;
 import static org.mule.runtime.internal.dsl.DslConstants.THIS_PREFIX;
-import static org.mule.runtime.module.extension.mule.internal.loader.parser.MuleSdkExtensionModelParser.APP_LOCAL_EXTENSION_NAMESPACE;
 import static org.mule.sdk.api.annotation.Extension.MULESOFT;
 
 import static java.util.Optional.empty;
 import static java.util.Optional.of;
 
+import org.mule.metadata.api.TypeLoader;
 import org.mule.runtime.api.meta.Category;
+import org.mule.runtime.ast.api.ArtifactAst;
+import org.mule.runtime.ast.api.ComponentAst;
+import org.mule.runtime.ast.internal.model.ExtensionModelHelper;
 import org.mule.runtime.module.extension.internal.loader.java.property.LicenseModelProperty;
+import org.mule.runtime.module.extension.internal.loader.parser.ExtensionModelParser;
 import org.mule.runtime.module.extension.internal.loader.parser.XmlDslConfiguration;
 
 import java.util.Optional;
+import java.util.stream.Stream;
 
 /**
- * {@link MuleSdkExtensionModelMetadataParser} implementation for Mule SDK extensions defined by applications.
+ * {@link ExtensionModelParser} implementation for Mule SDK in the context of applications.
  *
  * @since 4.5.0
  */
-public class MuleSdkApplicationExtensionModelMetadataParser implements MuleSdkExtensionModelMetadataParser {
+public class MuleSdkApplicationExtensionModelParser extends AbstractMuleSdkExtensionModelParser {
+
+  // The namespace of the extension when it's defined within an application rather than in a separate artifact.
+  public static final String APP_LOCAL_EXTENSION_NAMESPACE = "THIS";
 
   private final String extensionName;
 
-  public MuleSdkApplicationExtensionModelMetadataParser(String extensionName) {
+  public MuleSdkApplicationExtensionModelParser(String extensionName,
+                                                ArtifactAst ast,
+                                                TypeLoader typeLoader,
+                                                ExtensionModelHelper extensionModelHelper) {
+    super(ast, typeLoader, extensionModelHelper);
     this.extensionName = extensionName;
   }
 
@@ -62,5 +74,10 @@ public class MuleSdkApplicationExtensionModelMetadataParser implements MuleSdkEx
   @Override
   public LicenseModelProperty getLicenseModelProperty() {
     return new LicenseModelProperty(false, true, empty());
+  }
+
+  @Override
+  protected Stream<ComponentAst> getTopLevelElements(ArtifactAst ast) {
+    return ast.topLevelComponentsStream();
   }
 }
